+++ crates/bcr-ebill-web/src/handlers/bill.rs
@@ -12,10 +12,13 @@ use crate::data::{
 };
 use crate::service_context::ServiceContext;
 use bcr_ebill_api::data::bill::BillIssueData;
+use bcr_ebill_api::data::contact::{BillAnonParticipant, BillParticipant};
+use bcr_ebill_api::data::identity::IdentityType;
 use bcr_ebill_api::data::{
     bill::{BillAction, BillsFilterRole, LightBitcreditBillResult, RecourseReason},
-    contact::IdentityPublicData,
+    contact::BillIdentParticipant,
 };
+use bcr_ebill_api::service::Error;
 use bcr_ebill_api::service::bill_service::error::Error as BillServiceError;
 use bcr_ebill_api::util::file::{UploadFileHandler, detect_content_type_for_bytes};
 use bcr_ebill_api::util::{self, BcrKeys, ValidationError, currency};
@@ -36,19 +39,31 @@ pub async fn get_current_identity_node_id(state: &State<ServiceContext>) -> Stri
 
 pub async fn get_signer_public_data_and_keys(
     state: &State<ServiceContext>,
-) -> Result<(IdentityPublicData, BcrKeys)> {
+) -> Result<(BillParticipant, BcrKeys)> {
     let current_identity = state.get_current_identity().await;
     let local_node_id = current_identity.personal;
     let (signer_public_data, signer_keys) = match current_identity.company {
         None => {
             let identity = state.identity_service.get_full_identity().await?;
-            match IdentityPublicData::new(identity.identity) {
-                Some(identity_public_data) => (identity_public_data, identity.key_pair),
-                None => {
-                    return Err(
-                        service::Error::Validation(ValidationError::DrawerIsNotBillIssuer).into(),
-                    );
+            match identity.identity.t {
+                IdentityType::Ident => {
+                    match BillIdentParticipant::new(identity.identity) {
+                        Ok(identity_public_data) => (
+                            BillParticipant::Ident(identity_public_data),
+                            identity.key_pair,
+                        ),
+                        Err(_) => {
+                            // only non-anon bill issuers with a postal address can sign a bill
+                            return Err(
+                                Error::Validation(ValidationError::DrawerIsNotBillIssuer).into()
+                            );
+                        }
+                    }
                 }
+                IdentityType::Anon => (
+                    BillParticipant::Anon(BillAnonParticipant::new(identity.identity)),
+                    identity.key_pair,
+                ),
             }
         }
         Some(company_node_id) => {
@@ -63,7 +78,7 @@ pub async fn get_signer_public_data_and_keys(
                 .into());
             }
             (
-                IdentityPublicData::from(company),
+                BillParticipant::Ident(BillIdentParticipant::from(company)),
                 BcrKeys::from_private_key(&keys.private_key).map_err(service::Error::CryptoUtil)?,
             )
         }
@@ -369,6 +384,7 @@ pub async fn issue_bill(
             drawer_public_data: drawer_public_data.clone(),
             drawer_keys: drawer_keys.clone(),
             timestamp,
+            blank_issue: false,
         })
         .await?;
 
@@ -724,7 +740,16 @@ async fn request_recourse(
         .get_identity_by_node_id(recoursee_node_id)
         .await
     {
-        Ok(Some(buyer)) => buyer,
+        Ok(Some(BillParticipant::Ident(recoursee))) => recoursee,
+        Ok(Some(BillParticipant::Anon(_))) => {
+            // recoursee has to be identified
+            return Err(
+                BillServiceError::Validation(ValidationError::ContactIsAnonymous(
+                    recoursee_node_id.to_owned(),
+                ))
+                .into(),
+            );
+        }
         Ok(None) | Err(_) => {
             return Err(BillServiceError::RecourseeNotInContacts.into());
         }